invoke curve directly
