+++ src/link/index.js
@@ -1,9 +1,8 @@
+import {path} from "d3-path";
 import {slice} from "../array.js";
 import constant from "../constant.js";
-import line from "../line.js";
 import {x as pointX, y as pointY} from "../point.js";
-import pointRadial from "../pointRadial.js";
-import {bumpX as curveBumpX, bumpY as curveBumpY} from "../curve/bump.js";
+import {bumpX, bumpY, bumpRadial} from "../curve/bump.js";
 
 function linkSource(d) {
   return d.source;
@@ -14,20 +13,24 @@ function linkTarget(d) {
 }
 
 export function link(curve) {
-  var source = linkSource,
-      target = linkTarget,
-      x = pointX,
-      y = pointY,
-      l = line().curve(curve);
+  let source = linkSource;
+  let target = linkTarget;
+  let x = pointX;
+  let y = pointY;
+  let context = null;
+  let output = null;
 
   function link() {
+    let buffer;
     const argv = slice.call(arguments);
     const s = source.apply(this, argv);
     const t = target.apply(this, argv);
-    return l([
-      (argv[0] = s, [+x.apply(this, argv), +y.apply(this, argv)]),
-      (argv[0] = t, [+x.apply(this, argv), +y.apply(this, argv)])
-    ]);
+    if (context == null) output = curve(buffer = path());
+    output.lineStart();
+    argv[0] = s, output.point(+x.apply(this, argv), +y.apply(this, argv));
+    argv[0] = t, output.point(+x.apply(this, argv), +y.apply(this, argv));
+    output.lineEnd();
+    if (buffer) return output = null, buffer + "" || null;
   }
 
   link.source = function(_) {
@@ -47,50 +50,23 @@ export function link(curve) {
   };
 
   link.context = function(_) {
-    return arguments.length ? (l.context(_), link) : l.context();
+    return arguments.length ? (_ == null ? context = output = null : output = curve(context = _), link) : context;
   };
 
   return link;
 }
 
 export function linkHorizontal() {
-  return link(curveBumpX);
+  return link(bumpX);
 }
 
 export function linkVertical() {
-  return link(curveBumpY);
+  return link(bumpY);
 }
 
 export function linkRadial() {
-  var l = link(curveRadial);
+  const l = link(bumpRadial);
   l.angle = l.x, delete l.x;
   l.radius = l.y, delete l.y;
   return l;
-}
-
-class CurveRadial {
-  constructor(context) {
-    this._context = context;
-  }
-  lineStart() {
-    this._point = 0;
-  }
-  lineEnd() {}
-  point(x, y) {
-    x = +x, y = +y;
-    if (this._point++ === 0) {
-      this._x0 = x, this._y0 = y;
-    } else {
-      const p0 = pointRadial(this._x0, this._y0);
-      const p1 = pointRadial(this._x0, this._y0 = (this._y0 + y) / 2);
-      const p2 = pointRadial(x, this._y0);
-      const p3 = pointRadial(x, y);
-      this._context.moveTo(...p0);
-      this._context.bezierCurveTo(...p1, ...p2, ...p3);
-    }
-  }
-}
-
-function curveRadial(context) {
-  return new CurveRadial(context);
 }
